+++ packages/nextjs/app/intro/page.tsx
@@ -1,0 +1,3 @@
+export default function IntroPage() {
+  return <main>hello</main>;
+}
