+++ packages/nextjs/app/page.tsx
@@ -2,6 +2,7 @@
 
 import type { NextPage } from "next";
 import { useAccount } from "wagmi";
+import { BurnerWalletWrapper } from "~~/components/BurnerWalletWrapper";
 import { Header } from "~~/components/Header";
 import { History } from "~~/components/burnerwallet/History";
 import { useGetHistory } from "~~/hooks/useGetHistory";
@@ -11,7 +12,7 @@ const Home: NextPage = () => {
   const { chainId, isLoading, history, updateHistory } = useGetHistory({ address: connectedAddress });
 
   return (
-    <>
+    <BurnerWalletWrapper>
       <Header updateHistory={updateHistory} />
       <main>
         <div className="max-w-xl mx-auto">
@@ -20,7 +21,7 @@ const Home: NextPage = () => {
           </section>
         </div>
       </main>
-    </>
+    </BurnerWalletWrapper>
   );
 };
