+++ packages/nextjs/components/ScaffoldEthAppWithProviders.tsx
@@ -1,44 +1,16 @@
 "use client";
 
-import { useEffect } from "react";
 import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
-import { Toaster } from "react-hot-toast";
 import { WagmiConfig } from "wagmi";
-import { Footer } from "~~/components/Footer";
-import { QrCodeReader } from "~~/components/burnerwallet/QrCodeReader";
 import { BlockieAvatar } from "~~/components/scaffold-eth";
-import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
-import { useGlobalState } from "~~/services/store/store";
 import { wagmiConfig } from "~~/services/web3/wagmiConfig";
 import { appChains } from "~~/services/web3/wagmiConnectors";
-
-const ScaffoldEthApp = ({ children }: { children: React.ReactNode }) => {
-  const price = useNativeCurrencyPrice();
-  const setNativeCurrencyPrice = useGlobalState(state => state.setNativeCurrencyPrice);
-
-  useEffect(() => {
-    if (price > 0) {
-      setNativeCurrencyPrice(price);
-    }
-  }, [setNativeCurrencyPrice, price]);
-
-  return (
-    <>
-      <div className="max-w-xl mx-auto min-h-screen h-full bg-base-200 md:border-x border-base-100 shadow-lg">
-        {children}
-        <Footer />
-      </div>
-      <Toaster />
-      <QrCodeReader />
-    </>
-  );
-};
 
 export const ScaffoldEthAppWithProviders = ({ children }: { children: React.ReactNode }) => {
   return (
     <WagmiConfig config={wagmiConfig}>
       <RainbowKitProvider chains={appChains.chains} avatar={BlockieAvatar}>
-        <ScaffoldEthApp>{children}</ScaffoldEthApp>
+        {children}
       </RainbowKitProvider>
     </WagmiConfig>
   );
